refactor(model-access): move fact source into electron/shared/contracts (R26 boundary)
Integrating current origin/main brought the R26 import-boundary gate
(check:boundaries, PR #273). The 11 renderer imports of the model-access fact
source were flagged as new src→electron violations. Move the module to
electron/shared/contracts/modelAccessCapabilities.ts — the neutral contract
layer the boundary rule pre-exempts (pathNot '^electron/shared/contracts/') and
that R26 anticipates as the home renderer+main can both legally import.

- git mv electron/shared/modelAccessCapabilities.ts → electron/shared/contracts/
- update all 12 code importers + manifest.test + walk-stub comments + the
  vocabularies-baseline site anchor to the new path.

Verified: check:boundaries (0 new, 143 baseline), vocabularies, typecheck,
lint:ci (80/82), filesize, heavy-path, i18n, agents-sync; manifest.test 5/5.
