+++ src/ui/onboarding/AdapterVerificationScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { MODEL_ACCESS_ENTRY } from '../../../electron/shared/modelAccessCapabilities'
+import { MODEL_ACCESS_ENTRY } from '../../../electron/shared/contracts/modelAccessCapabilities'
 import { Group, Stack, Text } from '@mantine/core'
 import { IconAlertTriangle, IconCheck, IconRefresh } from '@tabler/icons-react'
 import { useTranslation } from 'react-i18next'
