+++ src/ui/onboarding/VendorOnboardCard.tsx
@@ -9,7 +9,7 @@
  * 样张：docs/design/mockups/onboarding-panel-A.html
  */
 import React from 'react'
-import { MODEL_ACCESS_ENTRY } from '../../../electron/shared/modelAccessCapabilities'
+import { MODEL_ACCESS_ENTRY } from '../../../electron/shared/contracts/modelAccessCapabilities'
 import { useTranslation } from 'react-i18next'
 import { IconKey, IconExternalLink } from '@tabler/icons-react'
 import { cn } from '../../utils/cn'
